backend/requirements.txt
Django==4.2.9
channels==4.0.0
channels-redis==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
daphne==4.0.0
redis==5.0.1
pytest==7.1.2
